# API/openapi.yaml
openapi: 3.0.3
info:
  title: UnlikeOtherAuthenticator API
  version: 1.0.0
  description: |
    Centralized OAuth / authentication service. Provides user auth flows,
    token issuance, organisation/team management, and admin operations.

    ## Authentication Schemes

    | Scheme | Header | Format | Used By |
    |--------|--------|--------|---------|
    | Domain Hash Bearer | `Authorization` | `Bearer SHA256(normalizedDomain+clientSecret)` | Domain-scoped server-to-server calls |
    | Admin Bearer | `Authorization` | `Bearer <HS256 JWT>` | Internal admin endpoints |
    | Access Token | `X-UOA-Access-Token` | `Bearer <JWT>` or raw JWT | User-context org endpoints |
    | Config JWT | `X-UOA-Config-JWT` | RS256 JWT | Email send endpoint |

    ## Config URL
    Most endpoints require a `config_url` query parameter pointing to a signed
    RS256 config JWT hosted by the integrator. The server fetches and verifies
    this JWT on every request.
servers:
  - url: https://auth.example.com
    description: Production
security: []
...
components:
  securitySchemes:
    domainHashBearer:
      type: http
      scheme: bearer
      description: |
        `Bearer SHA256(normalizedDomain + clientSecret)` — uses the per-domain client secret, not the shared secret.
    adminBearer:
      type: http
      scheme: bearer
      description: |
        HS256 JWT signed with `ADMIN_ACCESS_TOKEN_SECRET`. Claims must include
        `role: "superuser"` and `domain` matching `ADMIN_AUTH_DOMAIN`.
    accessToken:
      type: apiKey
      in: header
      name: X-UOA-Access-Token
      description: |
        UOA access token (HS256 JWT). May be prefixed with `Bearer ` or sent raw.
    configJwt:
      type: apiKey
      in: header
      name: X-UOA-Config-JWT
      description: RS256 config JWT signed with the domain's private key.
    billingAppKey:
      type: apiKey
      in: header
      name: X-UOA-App-Key
      description: One revocable UOA app key dedicated to the exact calling product/deployment and either entitlement or customer-lifecycle endpoints.
    billingActor:
      type: apiKey
      in: header
      name: X-UOA-Actor
      description: Short-lived RS256 actor assertion bound to the exact product app key and UOA user/org/team.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe endpoint signature verified over the exact raw request body with a separate webhook secret.
  parameters:
...
    domainQuery:
      name: domain
      in: query
      required: true
      description: Normalised domain (lowercase, trimmed).
      schema:
        type: string
        example: acme.example.com
...
    teamIdPath:
      name: teamId
      in: path
      required: true
      schema:
        type: string
...
    userIdPath:
      name: userId
      in: path
      required: true
      schema:
        type: string
...
    avatarStyleQuery:
      name: style
      in: query
      required: false
      description: |
        Overrides the generated-avatar style for this call. Only affects generated
        avatars — uploaded and proxied provider images are served as stored.
      schema:
        type: string
        enum: [tiles, waves, rings, mono]
    avatarSizeQuery:
      name: size
      in: query
      required: false
      description: |
        Rendered width/height of a generated SVG, clamped to 16-512 (default 128).
        The viewBox is constant; ignored for raster images.
      schema:
        type: integer
        minimum: 16
        maximum: 512
        default: 128
    avatarConfigUrlQuery:
      name: config_url
      in: query
      required: false
      description: |
        Optional HTTPS URL to the domain's signed config JWT. When supplied it is fetched
        and verified, and its `avatars.default_style` claim is used when `style` is absent.
        Its `domain` claim must match the authenticated domain.
      schema:
        type: string
        format: uri
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Generic public error message.
        code:
          type: string
          description: Machine-readable error code (only when DEBUG_ENABLED).
        summary:
          type: string
        details:
          type: array
          items:
            type: string
        hints:
          type: array
          items:
            type: string
...
    OkTrue:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
...
    AvatarUploadResponse:
      type: object
      required: [ok, avatar]
      properties:
        ok:
          type: boolean
          enum: [true]
        avatar:
          type: object
          required: [source, content_type, size_bytes, updated_at]
          properties:
            source:
              type: string
              enum: [uploaded]
            content_type:
              type: string
              enum: [image/png, image/jpeg, image/webp]
              description: The magic-byte sniffed type that was stored.
            size_bytes:
              type: integer
            updated_at:
              type: string
              format: date-time
    AvatarUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: |
            PNG, JPEG or WebP, at most 1 MiB. The stored type is decided by magic-byte
            sniffing, not the supplied mimetype; SVG and other non-raster uploads are
            rejected with the generic error envelope.
...
paths:
...
  /domain/users/{userId}/avatar:
    get:
      tags: [Avatars]
      summary: Get a user's avatar image
      operationId: getDomainUserAvatar
      description: |
        Returns image bytes, never JSON. Resolution precedence is fixed: uploaded image →
        server-side proxy of the provider `avatar_url` → deterministic generated SVG. Any
        provider-fetch failure falls back to the generated image, so a user visible to the
        authenticated domain always yields 200 with an image.

        The user must be visible to the domain under the same rules as `GET /domain/users`;
        unknown or cross-domain ids return the standard generic 404.
      security:
        - domainHashBearer: []
      parameters:
        - $ref: '#/components/parameters/userIdPath'
        - $ref: '#/components/parameters/domainQuery'
        - $ref: '#/components/parameters/avatarConfigUrlQuery'
        - $ref: '#/components/parameters/avatarStyleQuery'
        - $ref: '#/components/parameters/avatarSizeQuery'
      responses:
        '200':
          description: Avatar image bytes.
          headers:
            X-UOA-Avatar-Source:
              description: Which source the bytes came from.
              schema:
                type: string
                enum: [uploaded, provider, generated]
            Cache-Control:
              description: |
                `private, max-age=300` for uploaded/proxied images;
                `private, max-age=86400` for deterministic generated images.
              schema:
                type: string
            ETag:
              description: SHA-256 of the returned bytes.
              schema:
                type: string
            X-Content-Type-Options:
              schema:
                type: string
                enum: [nosniff]
            Content-Disposition:
              schema:
                type: string
            Content-Security-Policy:
              description: SVG responses only.
              schema:
                type: string
          content:
            image/svg+xml:
              schema:
                type: string
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '304':
          description: Not modified (matching `If-None-Match`).
        '401':
          description: Missing or invalid domain hash auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic not found — unknown or cross-domain user id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Avatars]
      summary: Set a user's uploaded avatar
      operationId: putDomainUserAvatar
      description: |
        Replaces the user's uploaded avatar. Rate-limited per domain + user.
        `users.avatar_url` is untouched, so a later social login cannot clobber the upload.
      security:
        - domainHashBearer: []
      parameters:
        - $ref: '#/components/parameters/userIdPath'
        - $ref: '#/components/parameters/domainQuery'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AvatarUploadRequest'
      responses:
        '200':
          description: Stored avatar metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUploadResponse'
        '400':
          description: Generic error — not a supported raster image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid domain hash auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Upload exceeds the 1 MiB cap.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Avatars]
      summary: Remove a user's uploaded avatar
      operationId: deleteDomainUserAvatar
      description: |
        Idempotent. Resolution falls back to the provider URL or the generated image.
        Rate-limited per domain + user.
      security:
        - domainHashBearer: []
      parameters:
        - $ref: '#/components/parameters/userIdPath'
        - $ref: '#/components/parameters/domainQuery'
      responses:
        '200':
          description: Removed.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
        '401':
          description: Missing or invalid domain hash auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /domain/teams/{teamId}/avatar:
    get:
      tags: [Avatars]
      summary: Get a team's (company) avatar image
      operationId: getDomainTeamAvatar
      description: |
        Returns image bytes, never JSON — the team counterpart of
        `/domain/users/{userId}/avatar`, for backend rendering with no end-user context.
        Resolution precedence is fixed: uploaded team image -> server-side proxy of the team
        `iconUrl` -> deterministic generated SVG seeded from the team id. Any proxy failure
        falls back to the generated image, so a team visible to the authenticated domain
        always yields 200 with an image.

        The team's organisation must belong to the authenticated domain; unknown or
        cross-domain team ids return the standard generic 404.
      security:
        - domainHashBearer: []
      parameters:
        - $ref: '#/components/parameters/teamIdPath'
        - $ref: '#/components/parameters/domainQuery'
        - $ref: '#/components/parameters/avatarConfigUrlQuery'
        - $ref: '#/components/parameters/avatarStyleQuery'
        - $ref: '#/components/parameters/avatarSizeQuery'
      responses:
        '200':
          description: Avatar image bytes.
          headers:
            X-UOA-Avatar-Source:
              description: Which source the bytes came from (`provider` = the proxied team `iconUrl`).
              schema:
                type: string
                enum: [uploaded, provider, generated]
            Cache-Control:
              description: |
                `private, max-age=300` for uploaded/proxied images;
                `private, max-age=86400` for deterministic generated images.
              schema:
                type: string
            ETag:
              description: SHA-256 of the returned bytes.
              schema:
                type: string
            X-Content-Type-Options:
              schema:
                type: string
                enum: [nosniff]
            Content-Disposition:
              schema:
                type: string
            Content-Security-Policy:
              description: SVG responses only.
              schema:
                type: string
          content:
            image/svg+xml:
              schema:
                type: string
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '304':
          description: Not modified (matching `If-None-Match`).
        '401':
          description: Missing or invalid domain hash auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic not found — unknown or cross-domain team id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Avatars]
      summary: Set a team's uploaded avatar (backend path)
      operationId: putDomainTeamAvatar
      description: |
        Replaces the team's uploaded avatar with no end-user context — the management path for
        product backends, which keep a bound refresh credential rather than a spendable UOA
        access token and therefore cannot drive the dual-auth `/org/*` mutations.

        No role check: per brief §24.10 the domain hash token represents full system trust for
        that domain, so the calling backend enforces its own owner/admin gating before relaying.
        The team's organisation must still belong to the authenticated domain. The team
        `iconUrl` column is untouched. Rate-limited per domain + team.
      security:
        - domainHashBearer: []
      parameters:
        - $ref: '#/components/parameters/teamIdPath'
        - $ref: '#/components/parameters/domainQuery'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AvatarUploadRequest'
      responses:
        '200':
          description: Stored avatar metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUploadResponse'
        '400':
          description: Generic error — not a supported raster image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid domain hash auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic not found — unknown or cross-domain team id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Upload exceeds the 1 MiB cap.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Avatars]
      summary: Remove a team's uploaded avatar (backend path)
      operationId: deleteDomainTeamAvatar
      description: |
        Idempotent. Resolution falls back to the team `iconUrl` or the generated image. Same
        full-trust domain-hash semantics as the PUT. Rate-limited per domain + team.
      security:
        - domainHashBearer: []
      parameters:
        - $ref: '#/components/parameters/teamIdPath'
        - $ref: '#/components/parameters/domainQuery'
      responses:
        '200':
          description: Removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkTrue'
        '401':
          description: Missing or invalid domain hash auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Generic not found — unknown or cross-domain team id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...